requirements.txt
flask
opencv-python
requests
numpy
python-dotenv
